requirements.txt
transformers
peft
bitsandbytes
accelerate
bitsandbytes-cuda117
rich
langchain_community
boto3
anthropic
langchain
langchain_anthropic
tabulate
uvicorn
streamlit==1.37.1
pandas
psycopg2
asyncpg
langchain_openai
langchain_experimental
streamlit
matplotlib
numpy
gtts
# playsound
google-cloud-texttospeech